Refactor: fix package name capitalization
