+++ shared/src/commonMain/kotlin/data/credentials/MobileDrivingLicenceCredentialAttributeCategorization.kt
@@ -1,7 +1,7 @@
 package data.credentials
 
 import at.asitplus.jsonpath.core.NormalizedJsonPath
-import at.asitplus.wallet.app.common.third_parts.at.asitplus.jsonpath.core.plus
+import at.asitplus.wallet.app.common.third_party.at.asitplus.jsonpath.core.plus
 import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements
 import data.PersonalDataCategory
