+++ shared/src/commonMain/kotlin/domain/BuildAuthenticationConsentPageFromAuthenticationRequestUriUseCase.kt
@@ -5,7 +5,7 @@ import at.asitplus.wallet.app.common.PresentationService
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import io.github.aakira.napier.Napier
 import kotlinx.serialization.encodeToString
-import ui.navigation.Routes.AuthenticationViewRoute
+import ui.navigation.routes.AuthenticationViewRoute
 
 class BuildAuthenticationConsentPageFromAuthenticationRequestUriUseCase(
     val presentationService: PresentationService,
