+++ shared/src/commonMain/kotlin/ui/navigation/WalletNavigation.kt
@@ -48,33 +48,33 @@ import kotlinx.serialization.json.Json
 import org.jetbrains.compose.resources.getString
 import ui.composables.BottomBar
 import ui.composables.NavigationData
-import ui.navigation.Routes.AddCredentialPreAuthnRoute
-import ui.navigation.Routes.AddCredentialRoute
-import ui.navigation.Routes.AuthenticationQrCodeScannerRoute
-import ui.navigation.Routes.AuthenticationSuccessRoute
-import ui.navigation.Routes.AuthenticationViewRoute
-import ui.navigation.Routes.CredentialDetailsRoute
-import ui.navigation.Routes.DCAPIAuthenticationConsentRoute
-import ui.navigation.Routes.ErrorRoute
-import ui.navigation.Routes.HomeScreenRoute
-import ui.navigation.Routes.LoadCredentialRoute
-import ui.navigation.Routes.LoadingRoute
-import ui.navigation.Routes.LogRoute
-import ui.navigation.Routes.OnboardingInformationRoute
-import ui.navigation.Routes.OnboardingStartRoute
-import ui.navigation.Routes.OnboardingTermsRoute
-import ui.navigation.Routes.OnboardingWrapperTestTags
-import ui.navigation.Routes.PreAuthQrCodeScannerRoute
-import ui.navigation.Routes.Route
-import ui.navigation.Routes.SettingsRoute
-import ui.navigation.Routes.SigningQtspSelectionRoute
-import ui.navigation.Routes.SigningRoute
+import ui.navigation.routes.AddCredentialPreAuthnRoute
+import ui.navigation.routes.AddCredentialRoute
+import ui.navigation.routes.AuthenticationQrCodeScannerRoute
+import ui.navigation.routes.AuthenticationSuccessRoute
+import ui.navigation.routes.AuthenticationViewRoute
+import ui.navigation.routes.CredentialDetailsRoute
+import ui.navigation.routes.DCAPIAuthenticationConsentRoute
+import ui.navigation.routes.ErrorRoute
+import ui.navigation.routes.HomeScreenRoute
+import ui.navigation.routes.LoadCredentialRoute
+import ui.navigation.routes.LoadingRoute
+import ui.navigation.routes.LogRoute
+import ui.navigation.routes.OnboardingInformationRoute
+import ui.navigation.routes.OnboardingStartRoute
+import ui.navigation.routes.OnboardingTermsRoute
+import ui.navigation.routes.OnboardingWrapperTestTags
+import ui.navigation.routes.PreAuthQrCodeScannerRoute
+import ui.navigation.routes.Route
+import ui.navigation.routes.SettingsRoute
+import ui.navigation.routes.SigningQtspSelectionRoute
+import ui.navigation.routes.SigningRoute
 import ui.screens.SelectIssuingServerView
 import ui.viewmodels.AddCredentialViewModel
-import ui.viewmodels.Authentication.AuthenticationQrCodeScannerViewModel
-import ui.viewmodels.Authentication.AuthenticationSuccessViewModel
-import ui.viewmodels.Authentication.DCAPIAuthenticationViewModel
-import ui.viewmodels.Authentication.DefaultAuthenticationViewModel
+import ui.viewmodels.authentication.AuthenticationQrCodeScannerViewModel
+import ui.viewmodels.authentication.AuthenticationSuccessViewModel
+import ui.viewmodels.authentication.DCAPIAuthenticationViewModel
+import ui.viewmodels.authentication.DefaultAuthenticationViewModel
 import ui.viewmodels.CredentialDetailsViewModel
 import ui.viewmodels.CredentialsViewModel
 import ui.viewmodels.LoadCredentialViewModel
@@ -83,9 +83,9 @@ import ui.viewmodels.PreAuthQrCodeScannerViewModel
 import ui.viewmodels.SettingsViewModel
 import ui.viewmodels.SigningQtspSelectionViewModel
 import ui.viewmodels.SigningViewModel
-import ui.views.Authentication.AuthenticationQrCodeScannerView
-import ui.views.Authentication.AuthenticationSuccessView
-import ui.views.Authentication.AuthenticationView
+import ui.views.authentication.AuthenticationQrCodeScannerView
+import ui.views.authentication.AuthenticationSuccessView
+import ui.views.authentication.AuthenticationView
 import ui.views.CredentialDetailsView
 import ui.views.CredentialsView
 import ui.views.ErrorView
