+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/AuthenticationConsentViewModel.kt
@@ -1,4 +1,4 @@
-package ui.viewmodels.Authentication
+package ui.viewmodels.authentication
 
 import androidx.compose.ui.graphics.ImageBitmap
 import at.asitplus.rqes.collection_entries.TransactionData
