+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/AuthenticationSuccessViewModel.kt
@@ -1,3 +1,3 @@
-package ui.viewmodels.Authentication
+package ui.viewmodels.authentication
 
 class AuthenticationSuccessViewModel(val navigateUp: () -> Unit)
