+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/AuthenticationViewModel.kt
@@ -1,4 +1,4 @@
-package ui.viewmodels.Authentication
+package ui.viewmodels.authentication
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
