+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/DCAPIAuthenticationViewModel.kt
@@ -1,4 +1,4 @@
-package ui.viewmodels.Authentication
+package ui.viewmodels.authentication
 
 import androidx.compose.ui.graphics.ImageBitmap
 import at.asitplus.dif.Constraint
